llm/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cohere==5.12.0
comm==0.2.2
cryptography==44.0.0
dataclasses-json==0.6.7
datasets==3.1.0
debugpy==1.8.9
decorator==5.1.1
deepeval==2.0
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
distro==1.9.0
docx2txt==0.8
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
faiss-cpu==1.9.0.post1
fastavro==1.9.7
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.63.2
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.3
idna==3.10
importlib-metadata==7.0.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.7.0
langchain==0.3.8
langchain-cohere==0.3.3
langchain-community==0.3.8
langchain-core==0.3.21
langchain-experimental==0.3.3
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langchain-upstage==0.4.0
langgraph==0.2.56
langgraph-checkpoint==2.0.8
langgraph-sdk==0.1.43
langsmith==0.1.146
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
openai==1.55.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pdfminer.six==20240706
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-hanspell==1.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
pykospacing @ git+https://github.com/haven-jeon/PyKoSpacing.git@b32a889cbd10b006d2f4aba118f0cd5b677e2979
pypdf==4.3.1
pysbd==0.3.4
pytest==8.3.3
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ragas==0.2.6
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.13.1
Send2Trash==1.8.3
sentry-sdk==2.19.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
soynlp==0.0.493
SQLAlchemy==2.0.35
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.4.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.1.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.14.0
types-python-dateutil==2.9.0.20241003
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
vectorstore==0.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0
